Flask/recipe-manager/requirements.txt
flask
flask_sqlalchemy
alembic
bcrypt
psycopg2-binary
flask_marshmallow
marshmallow-sqlalchemy
pytest
pytest-flask